requirements.txt
streamlit
pandas
ruamel.yaml
openai